+++ src/execute.rs
@@ -392,7 +392,7 @@ fn forward(
     let float_res: Vec<Vec<f32>> = res
         .iter()
         .zip(output_scales)
-        .map(|(t, scale)| t.iter().map(|e| (*e as f32 / scale)).collect_vec())
+        .map(|(t, scale)| t.iter().map(|e| ((*e as f64 / scale) as f32)).collect_vec())
         .collect();
     trace!("forward pass output: {:?}", float_res);
     data.output_data = float_res;
